Moved subclasses of VisibleFormItem to visible package
