+++ src/main/java/com/karankumar/bookproject/ui/book/components/form/item/visible/PageCount.java
@@ -1,4 +1,4 @@
-package com.karankumar.bookproject.ui.book.components.form.item;
+package com.karankumar.bookproject.ui.book.components.form.item.visible;
 
 import com.karankumar.bookproject.backend.entity.Book;
 import com.karankumar.bookproject.ui.book.BookFormErrors;
@@ -6,9 +6,9 @@ import com.vaadin.flow.component.textfield.IntegerField;
 import com.vaadin.flow.data.binder.Binder;
 import com.vaadin.flow.function.SerializablePredicate;
 
-public class SeriesPosition extends VisibleFormItem<IntegerField> {
+public class PageCount extends VisibleFormItem<IntegerField> {
 
-    public SeriesPosition() {
+    public PageCount(){
         super(new IntegerField());
     }
 
@@ -16,21 +16,22 @@ public class SeriesPosition extends VisibleFormItem<IntegerField> {
     public void configure() {
         IntegerField field = super.getField();
 
-        field.setPlaceholder("Enter series position");
+        field.setPlaceholder("Enter number of pages");
         field.setMin(1);
         field.setHasControls(true);
+        field.setClearButtonVisible(true);
     }
 
     @Override
     protected String getLabel() {
-        return "Series number";
+        return "Number of pages";
     }
 
     @Override
     public void bind(Binder<Book> binder, IntegerField fieldToCompare) {
         binder.forField(super.getField())
-              .withValidator(isNumberPositive(), BookFormErrors.SERIES_POSITION_ERROR)
-              .bind(Book::getSeriesPosition, Book::setSeriesPosition);
+              .withValidator(isNumberPositive(), BookFormErrors.PAGE_NUMBER_ERROR)
+              .bind(Book::getNumberOfPages, Book::setNumberOfPages);
     }
 
     private SerializablePredicate<Integer> isNumberPositive() {
