+++ src/main/java/com/karankumar/bookproject/ui/book/components/form/item/visible/ReadingStartDate.java
@@ -1,8 +1,7 @@
-package com.karankumar.bookproject.ui.book.components.form.item;
+package com.karankumar.bookproject.ui.book.components.form.item.visible;
 
 import com.karankumar.bookproject.backend.entity.Book;
 import com.karankumar.bookproject.ui.book.BookFormErrors;
-import com.karankumar.bookproject.ui.book.BookFormValidators;
 import com.vaadin.flow.component.datepicker.DatePicker;
 import com.vaadin.flow.data.binder.Binder;
 import com.vaadin.flow.function.SerializablePredicate;
